refactor QueryRunnerTestHelper
